# hoteli-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hoteli API Dokumentacioni",
        "description": "Dokumentacioni i API-së për Sistemin e Rezervimeve të Hotelit",
        "contact": {
            "email": "support@hoteli.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/rooms/search": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Kërko dhoma të disponueshme",
                "description": "Kërkon dhoma të disponueshme bazuar në kapacitetin dhe datat e check-in/check-out. Nuk kërkon autentifikim.",
                "operationId": "searchRooms",
                "parameters": [
                    {
                        "name": "capacity",
                        "in": "query",
                        "description": "Kapaciteti minimal i dhomës",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Data e check-in (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-09-10"
                        }
                    },
                    {
                        "name": "checkOutDate",
                        "in": "query",
                        "description": "Data e check-out (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-09-15"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista e dhomave të disponueshme",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Room"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Të dhëna të pavlefshme ose të mungojnë",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Të dhënat mungojnë"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Gabim serveri",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rooms/checkout": {
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Krijo një rezervim dhe përpuno pagesën",
                "description": "Përpunon një rezervim të ri dhe regjistron detajet e pagesës. Kërkon autentifikim.",
                "operationId": "processCheckout",
                "requestBody": {
                    "description": "Të dhënat e rezervimit dhe pagesës",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "room_id",
                                    "customer_name",
                                    "check_in",
                                    "check_out",
                                    "payment"
                                ],
                                "properties": {
                                    "room_id": {
                                        "description": "ID e dhomës që rezervohet",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "customer_name": {
                                        "description": "Emri i klientit që bën rezervimin",
                                        "type": "string",
                                        "example": "Besa Bota"
                                    },
                                    "check_in": {
                                        "description": "Data e mbërritjes (YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-09-10"
                                    },
                                    "check_out": {
                                        "description": "Data e largimit (YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-09-15"
                                    },
                                    "status": {
                                        "description": "Statusi i rezervimit (opsional, default 'confirmed')",
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "confirmed",
                                            "cancelled"
                                        ],
                                        "example": "confirmed"
                                    },
                                    "payment": {
                                        "required": [
                                            "cardholder",
                                            "bank_name",
                                            "card_number",
                                            "card_type",
                                            "cvv"
                                        ],
                                        "properties": {
                                            "cardholder": {
                                                "description": "Emri i mbajtësit të kartës",
                                                "type": "string",
                                                "example": "BESA BOTA"
                                            },
                                            "bank_name": {
                                                "description": "Emri i bankës",
                                                "type": "string",
                                                "example": "Bank Kombëtare"
                                            },
                                            "card_number": {
                                                "description": "Numri i kartës (16 shifra)",
                                                "type": "string",
                                                "pattern": "^[0-9]{16}$",
                                                "example": "1234567890123456"
                                            },
                                            "card_type": {
                                                "description": "Tipi i kartës",
                                                "type": "string",
                                                "enum": [
                                                    "Visa",
                                                    "MasterCard"
                                                ],
                                                "example": "Visa"
                                            },
                                            "cvv": {
                                                "description": "CVV (3 shifra)",
                                                "type": "string",
                                                "pattern": "^[0-9]{3}$",
                                                "example": "123"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rezervimi dhe pagesa u kryen me sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rezervimi dhe pagesa u kryen me sukses"
                                        },
                                        "reservation": {
                                            "$ref": "#/components/schemas/Reservation"
                                        },
                                        "payment": {
                                            "$ref": "#/components/schemas/Payment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Pa autentifikim"
                    },
                    "409": {
                        "description": "Konflikt rezervimi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Dhoma është e rezervuar për datat e zgjedhura"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Të dhëna të pavlefshme",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Gabim serveri",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Gabim gjatë përpunimit të rezervimit"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Mesazhi i gabimit"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Room": {
                "title": "Room",
                "description": "Modeli i dhomës së hotelit",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": 1
                    },
                    "room_number": {
                        "type": "string",
                        "example": "101"
                    },
                    "type": {
                        "type": "string",
                        "example": "Double"
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 2
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "available",
                            "occupied",
                            "dirty",
                            "maintenance"
                        ],
                        "example": "available"
                    },
                    "is_reserved": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
            "Reservation": {
                "title": "Reservation",
                "description": "Modeli i rezervimit",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": 1
                    },
                    "room_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "customer_name": {
                        "type": "string",
                        "example": "Jon Doe"
                    },
                    "check_in": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-09-10"
                    },
                    "check_out": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-09-15"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "confirmed",
                            "cancelled"
                        ],
                        "example": "confirmed"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
            "Payment": {
                "title": "Payment",
                "description": "Modeli i pagesës",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": 1
                    },
                    "reservation_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "cardholder": {
                        "type": "string",
                        "example": "JOHN DOE"
                    },
                    "bank_name": {
                        "type": "string",
                        "example": "MyBank"
                    },
                    "card_number": {
                        "description": "Numri i kartës i fshehur",
                        "type": "string",
                        "example": "XXXXXXXXXXXX1234"
                    },
                    "card_type": {
                        "type": "string",
                        "enum": [
                            "Visa",
                            "MasterCard"
                        ],
                        "example": "Visa"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 500
                    },
                    "paid_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-08 10:30:00"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
}